pyproject.toml
[project]
name = "ballerina-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard,standrad]>=0.115.7",
    "langchain>=0.3.16",
    "langchain-openai>=0.3.2",
    "langchain-pinecone>=0.2.2",
    "python-dotenv>=1.0.1",
]
